Move components to inside folders
