+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\git\\sing-along\\src\\index.tsx":"1","C:\\git\\sing-along\\src\\reportWebVitals.ts":"2","C:\\git\\sing-along\\src\\App.tsx":"3","C:\\git\\sing-along\\src\\hooks\\index.ts":"4","C:\\git\\sing-along\\src\\components\\index.ts":"5","C:\\git\\sing-along\\src\\hooks\\useFetch.tsx":"6","C:\\git\\sing-along\\src\\hooks\\useDimensions.tsx":"7","C:\\git\\sing-along\\src\\hooks\\useYouTube.tsx":"8","C:\\git\\sing-along\\src\\components\\Directory.tsx":"9","C:\\git\\sing-along\\src\\components\\Karaoke.tsx":"10","C:\\git\\sing-along\\src\\components\\SyncLyrics.tsx":"11"},{"size":581,"mtime":1639943682555,"results":"12","hashOfConfig":"13"},{"size":425,"mtime":1639943682555,"results":"14","hashOfConfig":"13"},{"size":768,"mtime":1639943682555,"results":"15","hashOfConfig":"13"},{"size":130,"mtime":1639943682555,"results":"16","hashOfConfig":"13"},{"size":120,"mtime":1639943682555,"results":"17","hashOfConfig":"13"},{"size":385,"mtime":1639943682555,"results":"18","hashOfConfig":"13"},{"size":739,"mtime":1639943682555,"results":"19","hashOfConfig":"13"},{"size":2895,"mtime":1639943682555,"results":"20","hashOfConfig":"13"},{"size":2998,"mtime":1639943682555,"results":"21","hashOfConfig":"13"},{"size":1775,"mtime":1639943682555,"results":"22","hashOfConfig":"13"},{"size":1369,"mtime":1639943682555,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"m8muf9",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\git\\sing-along\\src\\index.tsx",[],["47","48"],"C:\\git\\sing-along\\src\\reportWebVitals.ts",[],"C:\\git\\sing-along\\src\\App.tsx",[],"C:\\git\\sing-along\\src\\hooks\\index.ts",[],"C:\\git\\sing-along\\src\\components\\index.ts",[],"C:\\git\\sing-along\\src\\hooks\\useFetch.tsx",[],"C:\\git\\sing-along\\src\\hooks\\useDimensions.tsx",[],"C:\\git\\sing-along\\src\\hooks\\useYouTube.tsx",[],"C:\\git\\sing-along\\src\\components\\Directory.tsx",[],"C:\\git\\sing-along\\src\\components\\Karaoke.tsx",[],"C:\\git\\sing-along\\src\\components\\SyncLyrics.tsx",[],{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},"no-native-reassign",["53"],"no-negated-in-lhs",["54"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\git\\sing-along\\src\\index.tsx":"1","C:\\git\\sing-along\\src\\reportWebVitals.ts":"2","C:\\git\\sing-along\\src\\App.tsx":"3","C:\\git\\sing-along\\src\\hooks\\index.ts":"4","C:\\git\\sing-along\\src\\components\\index.ts":"5","C:\\git\\sing-along\\src\\hooks\\useFetch.tsx":"6","C:\\git\\sing-along\\src\\hooks\\useYouTube.tsx":"7","C:\\git\\sing-along\\src\\components\\SyncLyrics.tsx":"8","C:\\git\\sing-along\\src\\components\\Karaoke\\Karaoke.tsx":"9","C:\\git\\sing-along\\src\\components\\Karaoke\\useDimensions.tsx":"10","C:\\git\\sing-along\\src\\components\\Directory\\Directory.tsx":"11"},{"size":581,"mtime":1639943682555,"results":"12","hashOfConfig":"13"},{"size":425,"mtime":1639943682555,"results":"14","hashOfConfig":"13"},{"size":768,"mtime":1639943682555,"results":"15","hashOfConfig":"13"},{"size":81,"mtime":1639945317028,"results":"16","hashOfConfig":"13"},{"size":138,"mtime":1639945587703,"results":"17","hashOfConfig":"13"},{"size":385,"mtime":1639943682555,"results":"18","hashOfConfig":"13"},{"size":2895,"mtime":1639943682555,"results":"19","hashOfConfig":"13"},{"size":1369,"mtime":1639943682555,"results":"20","hashOfConfig":"13"},{"size":1818,"mtime":1639945343338,"results":"21","hashOfConfig":"13"},{"size":739,"mtime":1639945299615,"results":"22","hashOfConfig":"13"},{"size":2916,"mtime":1639945868969,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"m8muf9",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\git\\sing-along\\src\\index.tsx",[],["47","48"],"C:\\git\\sing-along\\src\\reportWebVitals.ts",[],"C:\\git\\sing-along\\src\\App.tsx",[],"C:\\git\\sing-along\\src\\hooks\\index.ts",[],"C:\\git\\sing-along\\src\\components\\index.ts",[],"C:\\git\\sing-along\\src\\hooks\\useFetch.tsx",[],"C:\\git\\sing-along\\src\\hooks\\useYouTube.tsx",[],"C:\\git\\sing-along\\src\\components\\SyncLyrics.tsx",[],"C:\\git\\sing-along\\src\\components\\Karaoke\\Karaoke.tsx",[],"C:\\git\\sing-along\\src\\components\\Karaoke\\useDimensions.tsx",[],"C:\\git\\sing-along\\src\\components\\Directory\\Directory.tsx",[],{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},"no-native-reassign",["53"],"no-negated-in-lhs",["54"],"no-global-assign","no-unsafe-negation"]
