+++ src/hooks/index.ts
@@ -1,3 +1,2 @@
-export { useDimensions } from './useDimensions';
 export { useFetch } from './useFetch';
 export { useYouTube } from './useYouTube';
